+++ CHANGELOG.md
@@ -1,13 +1,20 @@
-## 0.2.19 (??-??-2026)
+## 0.2.19 (22-06-2026)
+- Fixed a regression causing modern map exports to fail.
+- Fixed an issue that prevents certain WMO minimaps from exporting fully.
+
+## 0.2.18 (22-06-2026)
 - Added an "Export WMO Minimap" button to the model viewer for WMOs with minimaps.
 - Added "Maps" tab to MPQ Legacy builds with map previews and exports.
 - Added ability to export minimaps directly from maps for MPQ legacy builds.
 - Added ability to export terrain with pre-baked and alpha map support for MPQ legacy builds.
 - Added support for extended character customization (Dracthyr, Demon Hunter, Mechagnome etc).
 - Added support for gender/class item texture variants.
+- Added orthographic camera mode to model viewer.
+- Added JSON sidecar for minimap exports with world-space metadata.
+- Added dungeon journal encounter overlay on to map viewer.
+- Fixed terrain blending in the map viewer.
+- Fixed WMO minimaps to render properly for preview/export.
 - Fixed an issue that caused Mechagnome customization textures to render incorrectly.
-
-## 0.2.18 (11-05-2026)
 - Fixed GLTF primitives missing material index crashing Unreal Engine.
 - Improved rendering of WMO models in the map viewer.